pyproject.toml
[project]
name = "music-genre-classifier"
version = "1.0.0"
description = "High-performance deep learning system for music genre classification achieving 98.6% accuracy"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name", email = "your.email@example.com" }]
maintainers = [{ name = "Your Name", email = "your.email@example.com" }]
license = { text = "MIT" }
keywords = [
    "music",
    "genre",
    "classification",
    "deep-learning",
    "audio",
    "machine-learning",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "librosa>=0.11.0",
    "matplotlib>=3.8.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "plotly>=6.0.0",
    "scikit-learn>=1.3.0",
    "streamlit>=1.28.0",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "tqdm>=4.65.0",
    "umap-learn>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.4.0",
]
visualization = ["seaborn>=0.12.0", "shap>=0.42.0"]
api = ["flask>=2.3.0", "fastapi>=0.103.0", "uvicorn>=0.23.0"]
realtime = ["pyaudio>=0.2.11", "sounddevice>=0.4.6"]

[project.urls]
Homepage = "https://github.com/your-username/music-genre-classifier"
Documentation = "https://github.com/your-username/music-genre-classifier/tree/main/docs"
Repository = "https://github.com/your-username/music-genre-classifier"
"Bug Tracker" = "https://github.com/your-username/music-genre-classifier/issues"
Changelog = "https://github.com/your-username/music-genre-classifier/blob/main/CHANGELOG.md"

[project.scripts]
music-classifier = "src.app:main"
train-classifier = "scripts.train:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [".git", "__pycache__", ".venv", "build", "dist"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = ["tests/", "scripts/"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src --cov-report=term-missing"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[dependency-groups]
dev = ["pytest>=8.4.0", "pytest-cov>=6.2.1"]
